+++ src/cpx_io/cpx_system/cpx_ap/ap_supported_datatypes.py
@@ -5,7 +5,9 @@ SUPPORTED_IOL_DATATYPES = ["UINT8"]
 SUPPORTED_ISDU_DATATYPES = [
     "str",
     "int",
+    "uint",
+    "sint",
     "raw",
     "bool",
     "float",
-]  # TODO: add and/or test "sint"
+]