+++ tests/unit_tests/cpx_system/cpx_ap/test_ap_module.py
@@ -879,7 +879,7 @@ class TestApModule:
         module.write_channels([1, -2])
 
         # Assert
-        module.base.write_reg_data.assert_called_with(bytearray(b"\x01\xFE"), 0)
+        module.base.write_reg_data.assert_called_with(bytearray(b"\x01\xfe"), 0)
 
     def test_write_channels_uint8(self, module_fixture):
         """Test write_channels"""
@@ -971,7 +971,7 @@ class TestApModule:
         module.write_channels([1, -2])
 
         # Assert
-        module.base.write_reg_data.assert_called_with(bytearray(b"\x01\x00\xFE\xFF"), 0)
+        module.base.write_reg_data.assert_called_with(bytearray(b"\x01\x00\xfe\xff"), 0)
 
     def test_write_channels_uint16(self, module_fixture):
         """Test write_channels"""
@@ -2358,12 +2358,18 @@ class TestApModule:
         "input_value,, length, expected_output",
         [
             ("str", 3, b"str"),  # string
-            (1, 1, b"\x01"),  # int8
+            (1, 2, b"\x01\x00"),  # int8
             (0xCAFE, 2, b"\xfe\xca"),  # int16
             (0xBEBAFECA, 4, b"\xca\xfe\xba\xbe"),  # int32
+            (-1, 2, b"\xff\xff"),  # sint8
+            (-1925, 2, b"\x7b\xf8"),  # sint16
+            (-999999, 4, b"\xc1\xbd\xf0\xff"),  # 3byte sint32
+            (-99999999, 4, b"\x01\x1f\x0a\xfa"),  # sint32
             (b"\xca\xfe", 2, b"\xca\xfe"),  # bytes = raw
             (True, 1, b"\x01"),  # bool true
             (False, 1, b"\x00"),  # bool false
+            (0.0, 4, b"\x00\x00\x00\x00"),  # float 0
+            (-1.23456, 4, b"\xbf\x9e\x06\x10"),  # negative float
         ],
     )
     def test_write_isdu_different_datatypes(
@@ -2381,7 +2387,7 @@ class TestApModule:
         # Act
         module.write_isdu(input_value, 0, 0)
 
-        command = 51 if isinstance(input_value, (bool, int)) else 101
+        command = 51 if isinstance(input_value, (bool, int, float)) else 101
 
         # Assert
         module.base.write_reg_data.assert_has_calls(